pyproject.toml
[tool.poetry]
name = "test-blog-drf"
version = "0.1.0"
description = ""
authors = ["Максим Жданов <exodus.yarwood@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
